Source cleaning. wxMotif/OS2 adjustements.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@35778 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/motif/choice.cpp
@@ -352,7 +352,7 @@ void wxChoice::SetColumns(int n)
 {
     if (n<1) n = 1 ;
 
-    short numColumns = n ;
+    short numColumns = (short)n ;
     Arg args[3];
 
     XtSetArg(args[0], XmNnumColumns, numColumns);
